+++ packages/common/src/account/kmsAccount.ts
@@ -2,7 +2,7 @@ import { KMSClient } from "@aws-sdk/client-kms";
 import { LocalAccount, hashMessage, hashTypedData, keccak256, serializeTransaction } from "viem";
 import { toAccount } from "viem/accounts";
 import { getEthAddressFromKMS } from "./kms";
-import { createSignature } from "./eth";
+import { sign } from "./sign";
 
 export async function createKmsAccount(keyId: string, kmsInstance: KMSClient): Promise<LocalAccount> {
   const address = await getEthAddressFromKMS({ kmsInstance, keyId });
@@ -11,10 +11,10 @@ export async function createKmsAccount(keyId: string, kmsInstance: KMSClient): P
     address,
     async signMessage({ message }) {
       const hash = hashMessage(message);
-      const signature = await createSignature({
+      const signature = await sign({
         kmsInstance,
         keyId,
-        message: hash,
+        hash,
         address,
       });
 
@@ -23,10 +23,10 @@ export async function createKmsAccount(keyId: string, kmsInstance: KMSClient): P
     async signTransaction(transaction) {
       const unsignedTx = serializeTransaction(transaction);
       const hash = keccak256(unsignedTx);
-      const signature = await createSignature({
+      const signature = await sign({
         kmsInstance,
         keyId,
-        message: hash,
+        hash,
         address,
       });
 
@@ -34,10 +34,10 @@ export async function createKmsAccount(keyId: string, kmsInstance: KMSClient): P
     },
     async signTypedData(typedData) {
       const hash = hashTypedData(typedData);
-      const signature = await createSignature({
+      const signature = await sign({
         kmsInstance,
         keyId,
-        message: hash,
+        hash,
         address,
       });
