--- packages/common/src/account/types.ts
@@ -1,7 +1,0 @@
-import { KMSClient, SignCommandInput } from "@aws-sdk/client-kms";
-
-export type SignParams = {
-  keyId: SignCommandInput["KeyId"];
-  message: string;
-  kmsInstance: KMSClient;
-};
